toolkit/mozapps/update/tests/chrome/test_0085_error_patchApplyFailure_partial_complete_staging.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: error patching, download, and finished (partial failed and download complete), with fast MAR download"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

// We're going to simulate a very fast download by forcing it to complete
// before we click the "next" button on the initial wizard page. We do this
// by creating the continue file when the wizard loads to start the download,
// then clicking the button in the onStopRequest event listener.

const testDownloadListener = {
  onStartRequest(aRequest, aContext) { },

  onProgress(aRequest, aContext, aProgress, aMaxProgress) { },

  onStatus(aRequest, aContext, aStatus, aStatusText) { },

  onStopRequest(aRequest, aContext, aStatus) {
    debugDump("clicking next button");
    gDocElem.getButton("next").click();
    gAUS.removeDownloadListener(this);
  },

  QueryInterface: XPCOMUtils.generateQI([Ci.nsIRequestObserver,
                                         Ci.nsIProgressEventSink])
};

let TESTS = [ {
  pageid: PAGEID_ERROR_PATCHING,
  extraCheckFunction: createContinueFile
}, {
  pageid: PAGEID_DOWNLOADING,
  // Since the download is already finished, the hide button is disabled.
  buttonStates: {extra1: {disabled: true, hidden: true},
                   next: {disabled: false, hidden: false}}
}, {
  pageid: PAGEID_FINISHED,
  buttonClick: "extra1",
  extraStartFunction: removeContinueFile
} ];

// The buttons get configured differently on Mac, but this still works fine.
if (IS_MACOSX) {
  TESTS[1].buttonStates = {extra1: {disabled: true, hidden: true},
                             next: {disabled: false, hidden: true},
                           finish: {disabled: true, hidden: false}};
}

gUseTestUpdater = true;

function runTest() {
  debugDump("entering");

  removeContinueFile();

  let fastDownloadURL = URL_HTTP_UPDATE_XML + "?slowDownloadMar=1";
  let patches = getLocalPatchString("partial", null, null, null, null, null,
                                    STATE_PENDING) +
                getLocalPatchString("complete", fastDownloadURL, null, null,
                                    null, "false");
  // "0" for the backgroundInterval parameter forces a foreground download,
  // meaning downloading starts before the download wizard page is shown.
  let updates = getLocalUpdateString(patches, null, null, null,
                                     Services.appinfo.version, null,
                                     null, null, null, null, "false",
                                     null, null, null, null, null, "0");

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_FAILED_READ_ERROR);

  Services.prefs.setBoolPref(PREF_APP_UPDATE_STAGING_ENABLED, true);
  Services.prefs.setBoolPref(PREF_APP_UPDATE_SERVICE_ENABLED, false);

  reloadUpdateManagerData();

  testPostUpdateProcessing();

  gAUS.addDownloadListener(testDownloadListener);
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
